pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0,<69.3.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cotlette"
version = "0.0.0"
requires-python = ">= 3.6"
dependencies = [
    "uvicorn",
    "fastapi",
    "asgiref",
    "jinja2",
    "bcrypt",
    "python-jose",
]

authors = [
  {name = "Vladimir Penzin", email = "pvenv@icloud.com"},
]

description = "FastAPI sizzles, Django dazzles. The best of both worlds in one framework."

readme = "README.md"
license = {text = "MIT License"}
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.scripts]
cotlette = "cotlette.core.management:execute_from_command_line"

[project.urls]
Homepage = "https://github.com/ForceFledgling/cotlette"
Documentation = "https://github.com/ForceFledgling/cotlette"
Source = "https://github.com/ForceFledgling/cotlette"
Tracker = "https://github.com/ForceFledgling/cotlette/issues"

[tool.setuptools.packages.find]
where = ["src"]
